refactor(core): split agent_task_lifecycle god-file into module dir
